moved pages we have to proper locations, adjusted some things in gloabls.css in order to manage some color issues on some pages
