+++ hello-world/app/pages/Login.tsx
@@ -6,7 +6,7 @@ export default function Login() {
   const { signIn } = useAuth();
 
   return (
-    <div className="flex justify-center items-center h-screen bg-white">
+    <div className="flex justify-center items-center h-screen bg-custom-green">
       <div className="text-center">
         <h1 className="text-2xl font-bold mb-6 text-gray-800">
           Welcome to ShelfShare!
